ci: fix the clang, clang-tidy and sanitizer builds
clang rejects __builtin_cpu_supports("sha"), so SHA-NI is detected with cpuid. The DPDK stub
ignores the fields only the real implementation uses. The fixed-point test no longer adds past
INT64_MAX to show the plain operator wraps: that is undefined, and UBSan says so. The command-line
mains run under a guard that reports an escaping exception; the rest are clang-tidy's findings.

+++ tests/xdp/xdp_test_util.hpp
@@ -76,7 +76,7 @@ inline std::vector<std::byte> build_frame(const FrameSpec& s) {
     put16(f, 16, s.inner_ethertype);
   }
   // IPv4.
-  f[l3] = static_cast<std::byte>((s.version << 4U) | (s.ihl & 0x0FU));
+  f[l3] = static_cast<std::byte>((static_cast<unsigned>(s.version) << 4U) | (s.ihl & 0x0FU));
   const auto total = static_cast<int>(iph + udp_len) + s.tot_len_delta;
   put16(f, l3 + 2, static_cast<std::uint16_t>(total));
   put16(f, l3 + 4, 0x1234);
